requirements.txt
selenium
kafka-python
boto3
minio
apache-airflow
apache-airflow-providers-http
apache-airflow-providers-postgres
# ONNX and ML dependencies
onnxruntime
transformers
torch
numpy
pandas
# Spark dependencies
pyspark
# Additional ML dependencies
scikit-learn

